Fix parsing of type and specs
Type and specs code may need to be rewritten, that part is
strange and hackish, but now it parses the test examples I got.

The indentation is still way off what emacs does, I have not looked
at the indentation part at all, just fixed the parser errors.
But the status of emacs indentation of specs and types is not that
great either.