jsonrpc: Bugfix stdio LSP reader
There were two problems with the stdio reader:

1. io:get_line/1 will not work since (at least) lsp-mode sends data with
   a content-length that ends directly after the trailing "}" in the json
   structure, which means that io:get_line won't get that last "}" and
   try_decode never finds a complete message and the main loop "hangs".

   Perhaps other clients behave differently.

2. io:get_line/1 returns a string, whereas the main loop assumed a
   binary here:

        {data, Data} ->
                %?TRACE("STDIO", []),
                Buf2 = <<Buf/binary, Data/binary>>,

io:get_chars/2 would work, but forces us to read 1 character at
a time which is unnecessarily inefficient.

Open a port to read from standard input instead of using one of the io
built-ins.  Requires the erlang node to be started with parameter
"-noinput".